virtio: pci: fix 64-bit incompatibility
device_len is size_t, but written by pci_read_config_dword as a u32.
Change the type to fix this.

Link: https://lore.pengutronix.de/20210410104911.2079451-1-ahmad@a3f.at

+++ drivers/virtio/virtio_pci_common.h
@@ -47,7 +47,7 @@ struct virtio_pci_device {
 	void __iomem *notify_base;
 
 	/* So we can sanity-check accesses. */
-	size_t device_len;
+	u32 device_len;
 
 	/* Multiply queue_notify_off by this value. (non-legacy mode). */
 	u32 notify_offset_multiplier;
